+++ bin/host/Cargo.toml
@@ -11,43 +11,34 @@ publish = false
 workspace = true
 
 [dependencies]
-base-proof-mpt.workspace = true
-base-proof-executor.workspace = true
-base-proof-std-fpvm.workspace = true
-base-proof = { workspace = true, features = ["std"] }
-base-proof-preimage = { workspace = true, features = ["std"] }
-base-proof-driver.workspace = true
-base-consensus-derive.workspace = true
-base-consensus-registry.workspace = true
+revm = { workspace = true, features = ["std", "c-kzg", "secp256k1", "portable", "blst"] }
+rocksdb = { workspace = true, features = ["snappy", "bindgen-runtime"] }
 base-protocol = { workspace = true, features = ["std", "serde"] }
 base-consensus-genesis = { workspace = true, features = ["std", "serde"] }
-base-consensus-providers.workspace = true
 base-alloy-rpc-types-engine = { workspace = true, features = ["serde"] }
-base-alloy-network.workspace = true
-base-alloy-evm = { workspace = true, features = ["std"] }
-alloy-rlp.workspace = true
-alloy-transport.workspace = true
-alloy-eips = { workspace = true, features = ["kzg"] }
-alloy-serde.workspace = true
-alloy-provider = { workspace = true, features = ["reqwest"] }
-alloy-consensus.workspace = true
-alloy-rpc-client.workspace = true
-alloy-transport-http.workspace = true
+base-proof = { workspace = true, features = ["std"] }
 alloy-rpc-types = { workspace = true, features = ["eth", "debug"] }
 alloy-primitives = { workspace = true, features = ["serde"] }
+alloy-eips = { workspace = true, features = ["kzg"] }
+tracing-subscriber = { workspace = true, features = ["fmt", "env-filter"] }
+serde = { workspace = true, features = ["derive"] }
+clap = { workspace = true, features = ["derive", "env"] }
+tokio = { workspace = true, features = ["full"] }
+base-alloy-network.workspace = true
+base-consensus-providers.workspace = true
+base-proof-host.workspace = true
+base-proof-preimage.workspace = true
+base-proof-std-fpvm.workspace = true
 alloy-rpc-types-beacon.workspace = true
+alloy-consensus.workspace = true
+alloy-provider.workspace = true
 alloy-network.workspace = true
-revm = { workspace = true, features = ["std", "c-kzg", "secp256k1", "portable", "blst"] }
-anyhow.workspace = true
 tracing.workspace = true
 serde_json.workspace = true
 async-trait.workspace = true
-rocksdb = { workspace = true, features = ["snappy", "bindgen-runtime"] }
-tokio = { workspace = true, features = ["full"] }
-serde = { workspace = true, features = ["derive"] }
-clap = { workspace = true, features = ["derive", "env"] }
-tracing-subscriber = { workspace = true, features = ["fmt", "env-filter"] }
 thiserror.workspace = true
+alloy-rlp.workspace = true
+anyhow.workspace = true
 ark-ff.workspace = true
 
 [dev-dependencies]
