+++ crates/proof/host/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "base-proof-host"
+version = "0.0.0"
+edition.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+publish = false
+
+[lints]
+workspace = true
+
+[dependencies]
+anyhow.workspace = true
+async-trait.workspace = true
+thiserror.workspace = true
+tokio.workspace = true
+tracing.workspace = true
+alloy-consensus.workspace = true
+alloy-primitives.workspace = true
+alloy-provider.workspace = true
+alloy-rlp.workspace = true
+alloy-transport.workspace = true
+serde_json.workspace = true
+base-proof-mpt.workspace = true
+base-proof-preimage.workspace = true
+base-proof = { workspace = true, features = ["std"] }
