+++ crates/proof/host/README.md
@@ -1,0 +1,3 @@
+# base-proof-host
+
+Generic host-side infrastructure for the Base fault proof host.
