+++ crates/proof/host/src/backend/offline.rs
