+++ crates/proof/host/src/kv/mem.rs
